Initial implementation of catboost access in library-bridge
This commit moves the catboost model evaluation out of the server
process into the library-bridge binary. This serves two goals: On the
one hand, crashes / memory corruptions of the catboost library no longer
affect the server. On the other hand, we can forbid loading dynamic
libraries in the server (catboost was the last consumer of this
functionality), thus improving security.

SQL syntax:

  SELECT
    catboostEvaluate('/path/to/model.bin', FEAT_1, ..., FEAT_N) > 0 AS prediction,
    ACTION AS target
  FROM amazon_train
  LIMIT 10

Required configuration:

  <catboost_lib_path>/path/to/libcatboostmodel.so</catboost_lib_path>

*** Implementation Details ***

The internal protocol between the server and the library-bridge is
simple:

- HTTP GET on path "/extdict_ping":
  A ping, used during the handshake to check if the library-bridge runs.

- HTTP POST on path "extdict_request"
  - Method "castboost_libNew": load libcatboost into the library-bridge.
    To allow parallel usage, the library-bridge keeps a separate state
    (model handle) for each path to a compiled model.
  - Methods "catboost_GetTreeCount" and "castboost_libEvaluate": send
    inputs (float and categorical features), then evaluate the model in
    library-bridge, then send results back to server.

Once loaded, libcatboost.so is never unloaded from the library-bridge,
i.e. to use a libcatboost.so in a different location, the server and the
library-bridge must be restarted. This makes sense
- as the same library instance is able to evaluate different models,
- a shared library load is rather expensive (400 ms on my debug build),
- and it avoids the complexity of having to send the library path and
  check the library load status for every request.

To allow the evaluation of the same model by different users/sessions,
the model state in the library-bridge is protected by a mutex.